requirements.txt
pillow==12.1.1
Django==6.0.1
django-ckeditor==6.7.3
PyMySQL==1.1.2

